fix(clerk-js): Allow toggling default phones only if totp id off to avoid confusion
